Show reminder that hardware may need some interaction
For example when setting up Yubikey for SSH with physical presence
proof[1] every signature (every SSH connection) needs the yubikey to
be touched.

The blinking LED indicating this can be easy to miss, especially if
it's a Nano model stuffed into the side of a laptop or out of sight in
a workstation on the floor.

This patch is a proof of concept, and while it shouldn't have
downsides where no interaction is required, unless the smartcard takes >100ms
to perform a signature, it should probably be much more configurable.

[1] https://blog.habets.se/2016/01/Yubikey-4-for-SSH-with-physical-presence-proof.html